+++ src/wanhive-base.h
@@ -18,9 +18,8 @@
  */
 #include "base/common/Activity.h"
 #include "base/common/CommandLine.h"
+#include "base/common/Duplex.h"
 #include "base/common/Logger.h"
-#include "base/common/Sink.h"
-#include "base/common/Source.h"
 
 /*
  * Databases