+++ docs/ChangeLog.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Base class for bidirectional (**Duplex**) channels.
+
+### Changed
+
+- Stream Watcher extends **Duplex**, not **Sink** or **Source**.
+- The callback function for byte streams in **Hub** accepts the stream's channel as its parameter.
+
 ## [16.0.0] - 2025-12-09
 
 ### Added